make tags optional
In certain cases it can be useful to have the ability to not generated
tags, so this adds a flag which makes it possible to disable the
generation of flags.

Furthermore it also uses `name` for the `.Metric` attribute, instead of
using the previous "some.id.of.a.metric". That's necessary because the
`.SetId()` method generates the ID based on the `.Metric` attribute and the
tags, so if there are no tags set then all generated metrics will end up
with the same ID.